refactor(tokens): create web.cjs config
